Revert ".."
This reverts commit 055a95ba8744470185f01cfbdbf3d5b1c087d881.

+++ application/controllers/Datasiswa.php
@@ -1,6 +1,6 @@
 <?php
 
-class Datasiswa extends CI_Controller
+class Multiple extends CI_Controller
 {
     public function __construct()
     {
@@ -8,6 +8,7 @@ class Datasiswa extends CI_Controller
         $this->load->model('Mcrud');
         $this->load->library('form_validation');
     }
+
     public function index()
     {
         $this->data['title_web'] = 'Data Siswa ';
@@ -17,7 +18,9 @@ class Datasiswa extends CI_Controller
         $this->load->view('sidebar_view', $this->data);
         $this->load->view('siswa/index', $this->data);
         $this->load->view('footer_view', $this->data);
+        // $this->load->view('multiple/index', $this->data);
     }
+
     public function add()
     {
         $this->form_validation->set_rules(
@@ -49,7 +52,7 @@ class Datasiswa extends CI_Controller
             ];
             $this->db->insert('users', $data);
             $this->session->set_flashdata('users', 'Berhasil ditambahkan!');
-            redirect('Datasiswa');
+            redirect('Multiple');
         }
     }
     public function edit_siswa($id)
@@ -101,15 +104,33 @@ class Datasiswa extends CI_Controller
             // var_dump($this->db->last_query());
             // die;
             $this->session->set_flashdata('users', 'telah diganti');
-            redirect('Datasiswa');
+            redirect('multiple/index');
         }
     }
-    public function delete()
+    public function delete($id)
     {
-
-        $id = htmlspecialchars($this->input->post('id'));
-        $this->db->where('id', $id);
-        $this->db->delete('users');
-        redirect('Datasiswa');
+        if ($id == NULL) {
+            $this->db->get_where('users', ['id' => $id])->result_array();
+        } else {
+            return $this->db->get_where('users', ['id' => $id])->row_array();
+            $id = htmlspecialchars($this->input->post('id'));
+            $this->db->where('id', $id);
+            $this->db->delete('users');
+            $this->sesion->set_flasdata('users', 'Telah Dihapus');
+            redirect('Multiple');
+        }
     }
-}
+    public function edit($id)
+    {
+        // $data['kode_area'] = $this->db->get_where('kode_area', ['id_user'=>$id])->result();
+        $this->data['title_web'] = 'Data Siswa ';
+        $this->data['idbo'] = $this->session->userdata('ses_id');
+        $this->data['users'] = $this->db->get('users')->result();
+        $this->data['kode_area'] = $this->db->get_where('kode_area', ['id' => $id])->result_array();
+        $this->data['get_id'] = $id;
+        $this->load->view('header_view', $this->data);
+        $this->load->view('sidebar_view', $this->data);
+        $this->load->view('siswa/edit', $this->data);
+        $this->load->view('footer_view', $this->data);
+        // $this->load->view('multiple/edit', $data);
+    }ssss
